Fix build if userid does not exist in user database
make_build_data uses the output of the id command in order to generate the
string in build_data.c indicating who built the program. Unfortunately, the
previous parsing code looked for the presence of a ")" in order to determine
where the user field ended. If the user is not found in the user database,
the output looks more like:
uid=59784003 gid=60243 groups=60243

Instead, now it just calls the glibc helpers and formats the line based
upon those instead of shelling out and trying to parse the output of id.
If it cannot retrieve the username, it just sets the UID.

Signed-off-by: Sargun Dhillon <sargun@sargun.me>